requirements.txt
asgiref==3.4.1
crispy-tailwind==0.5.0
Django==3.2.19
django-crispy-forms==1.13.0
django-debug-toolbar==3.7.0
django-environ==0.8.1
django-extensions==3.2.1
django-tables2==2.4.1
Pillow==9.3.0
pydotplus==2.0.2
pyparsing==3.0.9
pytailwindcss==0.1.2
pytz==2021.3
sqlparse==0.4.4
psycopg2-binary==2.9.1
https://github.com/benoitc/gunicorn/archive/refs/heads/master.zip#egg=gunicorn==20.1.0
https://github.com/alexdeathway/django-summernote/archive/refs/tags/patch.zip
eventlet==0.33.3
dnspython==2.3.0
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2

#[api-packages]
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.10.0